seperate navigators into components
